revert changes and also added popup on click of chat + button
