+++ ClientAdvisor/App/frontend/src/components/Cards/Cards.module.css
@@ -49,51 +49,3 @@
     font-weight: bold;
     margin-bottom: 8px;
   } */
-
-  /* popup css*/
-  .popupContainer {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background: #fff;
-    border: 1px solid #ccc;
-    padding: 10px;
-    border-radius: 5px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-    z-index: 1;
-    position: fixed;
-    right: 1rem;
-    top: 6rem;
-  }
-
-  .clientName{
-    color: #4CAF50;
-  }
-
-  .popupContent {
-    display: flex;
-    align-items: center;
-  }
-  
-  .checkmark {
-    color: white;
-    font-size: 13px;
-    background: #4caf50;
-  }
-  
-  .popupText {
-    margin-left: 10px;
-  }
-  
-  .popupSubtext {
-    font-size: smaller;
-    color: #666;
-  }
-  
-  .closeButton {
-    background: none;
-    border: none;
-    font-size: 16px;
-    cursor: pointer;
-  }
-
