+++ ClientAdvisor/App/frontend/src/pages/chat/Chat.module.css
@@ -370,3 +370,58 @@ a {
     max-width: 100%;
   }
 }
+
+
+ /* popup css*/
+ .popupContainer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background: #fff;
+  border: 1px solid #ccc;
+  padding: 10px;
+  border-radius: 5px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  z-index: 1;
+  position: fixed;
+  right: 1rem;
+  top: 6rem;
+}
+
+.clientName{
+  color: #4CAF50;
+}
+
+.popupContent {
+  display: flex;
+  align-items: center;
+}
+
+.checkmark {
+  color: white;
+  font-size: 13px;
+  vertical-align: middle;
+  padding-right: 0.4rem;
+}
+
+.popupText {
+  margin-left: 10px;
+}
+
+
+.popupSubtext {
+  font-size: smaller;
+  color: #666;
+  padding-left: 1.4rem;
+  width: 75%;
+
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  font-size: 13px;
+  cursor: pointer;
+  position: absolute;
+  right: 0.8rem;
+}
