+++ ClientAdvisor/App/frontend/src/pages/chat/Chat.tsx
@@ -11,6 +11,7 @@ import DOMPurify from 'dompurify'
 
 import styles from './Chat.module.css'
 import TeamAvatar from '../../assets/TeamAvatar.svg'
+import TickIcon  from '../../assets/TickIcon.svg'
 import { XSSAllowTags } from '../../constants/xssAllowTags'
 
 import {
@@ -57,6 +58,7 @@ const Chat = () => {
   const [clearingChat, setClearingChat] = useState<boolean>(false)
   const [hideErrorDialog, { toggle: toggleErrorDialog }] = useBoolean(true)
   const [errorMsg, setErrorMsg] = useState<ErrorMessage | null>()
+  const [isVisible, setIsVisible] = useState(false);
 
   const errorDialogContentProps = {
     type: DialogType.close,
@@ -97,6 +99,19 @@ const Chat = () => {
       setErrorMsg(null)
     }, 500)
   }
+  const closePopup = () => {
+    setIsVisible(!isVisible);
+  };
+
+  useEffect(() => {
+    if (isVisible) {
+      const timer = setTimeout(() => {
+        setIsVisible(false);
+      }, 2000); // Popup will disappear after 3 seconds
+
+      return () => clearTimeout(timer); // Cleanup the timer on component unmount
+    }
+  }, [isVisible]);
 
   useEffect(() => {
     setIsLoading(appStateContext?.state.chatHistoryLoadingState === ChatHistoryLoadingState.Loading)
@@ -582,6 +597,9 @@ const Chat = () => {
   }
 
   const newChat = () => {
+     
+    console.log("new chacht onlick")
+    setIsVisible(true);
     setProcessMessages(messageStatus.Processing)
     setMessages([])
     setIsCitationPanelOpen(false)
@@ -701,6 +719,18 @@ const Chat = () => {
 
   return (
     <div className={styles.container} role="main">
+      {isVisible && (
+        <div className={styles.popupContainer}>
+          <div className={styles.popupContent}>
+            <div className={styles.popupText}>
+              <div><span className={styles.checkmark}><img src={TickIcon} /></span>Chat saved
+              <button className={styles.closeButton} onClick={closePopup}>X</button></div>
+              <div className={styles.popupSubtext}><span className={styles.popupMsg}>Your chat history has been saved successfully!</span></div>
+            </div>
+            
+          </div>
+        </div>
+    )}
       {showAuthMessage ? (
         <Stack className={styles.chatEmptyState}>
           <ShieldLockRegular
